moved dependency helpers to utils.py
