+++ aws_lambda_builders/workflows/nodejs_npm/actions.py
@@ -2,12 +2,11 @@
 Action to resolve NodeJS dependencies using NPM
 """
 
-import json
 import logging
-from os import sep as path_sep
 
 from aws_lambda_builders.actions import BaseAction, Purpose, ActionFailedError
 from .npm import NpmExecutionError
+from .utils import DependencyUtils
 
 LOG = logging.getLogger(__name__)
 LOG.setLevel(logging.DEBUG)
@@ -72,16 +71,16 @@ class NodejsNpmPackAction(BaseAction):
             LOG.debug("NODEJS searching for local dependencies")
 
             local_manifest_path = self.osutils.joinpath(self.artifacts_dir, 'package', 'package.json')
-            local_dependencies = get_local_dependencies(self.manifest_path)
+            local_dependencies = DependencyUtils.get_local_dependencies(self.manifest_path)
             for (dep_name, dep_path) in local_dependencies.items():
                 dep_scratch_dir = self.osutils.joinpath(self.scratch_dir, str(abs(hash(dep_name))))
                 dep_artifacts_dir = self.osutils.joinpath(dep_scratch_dir, 'unpacked')
-                dependency_tarfile_path = package_local_dependency(package_path[5:], dep_path, dep_artifacts_dir, dep_scratch_dir, self.osutils, self.subprocess_npm)
+                dependency_tarfile_path = DependencyUtils.package_local_dependency(package_path[5:], dep_path, dep_artifacts_dir, dep_scratch_dir, self.osutils, self.subprocess_npm)
                 local_packaged_dep_path = self.osutils.joinpath(self.artifacts_dir, 'package', 'local_dep')
                 if not self.osutils.dir_exists(local_packaged_dep_path):
                     self.osutils.mkdir(local_packaged_dep_path)
                 dependency_tarfile_path = self.osutils.copy_file(dependency_tarfile_path, self.osutils.joinpath(local_packaged_dep_path))
-                update_manifest(local_manifest_path, dep_name, dependency_tarfile_path, self.osutils)
+                DependencyUtils.update_manifest(local_manifest_path, dep_name, dependency_tarfile_path, self.osutils)
 
         except NpmExecutionError as ex:
             raise ActionFailedError(str(ex))
@@ -213,89 +212,3 @@ class NodejsNpmrcCleanUpAction(BaseAction):
 
         except OSError as ex:
             raise ActionFailedError(str(ex))
-
-
-def get_local_dependencies(manifest_path):
-    """
-    Helper function to extract all local dependencies in a package.json manifest
-    """
-
-    with open(manifest_path) as manifest_file:
-        manifest = json.loads(manifest_file.read())
-
-        if 'dependencies' in manifest:
-            return dict((k, v) for (k, v) in manifest['dependencies'].items() if is_local_dependency(v))
-        else:
-            return {}
-
-
-def is_local_dependency(path):
-    """
-    Helper function to check if package dependency is a local package
-    """
-
-    try:
-        return path.startswith('file:') or path.startswith('.')
-    except AttributeError:
-        return False
-
-
-def package_local_dependency(parent_package_path, rel_package_path, artifact_dir, scratch_dir, osutils, subprocess_npm):
-    if rel_package_path.startswith('file:'):
-        rel_package_path = rel_package_path[5:].strip()
-
-    if rel_package_path.startswith('.'):
-        package_path = osutils.abspath(osutils.joinpath(parent_package_path, rel_package_path))
-
-    LOG.debug("NODEJS packaging dependency %s to %s", package_path, scratch_dir)
-
-    if not osutils.dir_exists(scratch_dir):
-        osutils.mkdir(scratch_dir)
-
-    tarfile_name = subprocess_npm.run(["pack", "-q", package_path], cwd=scratch_dir).splitlines()[-1]
-
-    LOG.debug("NODEJS packed dependency to %s", tarfile_name)
-
-    tarfile_path = osutils.joinpath(scratch_dir, tarfile_name)
-
-    LOG.debug("NODEJS extracting to %s", artifact_dir)
-
-    osutils.extract_tarfile(tarfile_path, artifact_dir)
-
-    LOG.debug("NODEJS searching for subpackage local dependencies")
-
-    # manifest_path = osutils.joinpath(package_path, 'package.json')
-    # local_dependencies = get_local_dependencies(manifest_path)
-    # for (dep_name, dep_path) in local_dependencies.items():
-    #     dep_scratch_dir = osutils.joinpath(scratch_dir, '..', hash(dep_name))
-    #     dep_artifacts_dir = osutils.joinpath(dep_scratch_dir, 'unpacked')
-    #     dependency_tarfile_path = package_local_dependency(package_path[5:], dep_path, dep_artifacts_dir, dep_scratch_dir, self.osutils, self.subprocess_npm)
-    #     update_manifest(manifest_path, dep_name, dependency_tarfile_path)
-
-    localized_package_dir = osutils.joinpath(artifact_dir, 'package')
-
-    LOG.debug("NODEJS repackaging dependency %s to %s", artifact_dir, localized_package_dir)
-
-    tarfile_name = subprocess_npm.run(["pack", "-q", localized_package_dir], cwd=localized_package_dir).splitlines()[-1]
-
-    return osutils.joinpath(localized_package_dir, tarfile_name)
-
-
-def update_manifest(manifest_path, dep_name, dependency_tarfile_path, osutils):
-    """
-    Helper function to update dependency path to localized tar
-    """
-
-    package_path = osutils.dirname(manifest_path)
-    manifest_backup = osutils.copy_file(manifest_path, f'{manifest_path}.bak')
-
-    with open(manifest_backup, 'r') as manifest_backup_file:
-        manifest = json.loads(manifest_backup_file.read())
-
-        if 'dependencies' in manifest and dep_name in manifest['dependencies']:
-            dep_rel_path = osutils.relative_path(dependency_tarfile_path, start=package_path)
-            dep_rel_path = osutils.joinpath('.', dep_rel_path).replace(path_sep, "/")
-            manifest['dependencies'][dep_name] = f'file:{dep_rel_path}'
-
-            with open(manifest_path, 'w') as manifest_write_file:
-                manifest_write_file.write(json.dumps(manifest, indent=4))
